[lk][trusty] Update structure member len form size_t to uint32_t
Update structure member len in struct iover_kern_t, iovec_user_t
and ipc_msg_info form size_t to uint32_t

+++ lib/trusty/include/lib/trusty/ipc_msg.h
@@ -64,7 +64,7 @@ typedef struct ipc_msg_user {
 } ipc_msg_user_t;
 
 typedef struct ipc_msg_info {
-	size_t	len;
+	uint32_t	len;
 	uint32_t	id;
 } ipc_msg_info_t;
